refactor logic, use more throws
