+++ examples/oink-oink/src/lib/getRandomItem.ts
@@ -1,0 +1,3 @@
+export function getRandomItem<T>(items: readonly T[]) {
+  return items[Math.floor(Math.random() * items.length)]
+}
