+++ examples/oink-oink/src/lib/isLastActor.ts
@@ -1,0 +1,6 @@
+import { GameState } from "./types/GameState"
+
+export function isLastActor(game: GameState) {
+  const actorIndex = game.players.findIndex((player) => player.actor)
+  return actorIndex + 1 === game.players.length
+}
