+++ examples/oink-oink/src/logic.ts
@@ -1,4 +1,9 @@
-import { GameState, animals, emotions } from "./lib/types/GameState"
+import { animals, emotions } from "./lib/types/GameState"
+import { getRandomItem } from "./lib/getRandomItem"
+import { setActor } from "./lib/setActor"
+import { isLastActor } from "./lib/isLastActor"
+import { newTurn } from "./lib/newTurn"
+import { startGameCheck } from "./lib/startGameCheck"
 
 export const numRounds = 3
 export const turnCountdown = 3
@@ -10,8 +15,6 @@ export const turnAlmostOverAt = 5
 export const endOfTurnDuration = 3
 
 export const displayCorrectGuessFor = 3
-
-// todo: replace all ifs with asserts/throws?
 
 Rune.initLogic({
   minPlayers: 3,
@@ -34,12 +37,16 @@ Rune.initLogic({
   actions: {
     setReadyToStart: (_, { game, playerId }) => {
       if (game.gameStarted) throw Rune.invalidAction()
+
       const player = game.players.find((player) => player.id === playerId)
-      if (player) player.readyToStart = true
+
+      if (!player) throw Rune.invalidAction()
+      player.readyToStart = true
+
       startGameCheck(game)
     },
     makeGuess: ({ animal, emotion }, { game, playerId }) => {
-      if (!game.currentTurn) return
+      if (!game.currentTurn) throw Rune.invalidAction()
 
       const guess = {
         playerId,
@@ -56,7 +63,7 @@ Rune.initLogic({
         const player = game.players.find((player) => player.id === playerId)
         const actor = game.players.find((player) => player.actor)
 
-        if (!player || !actor) return
+        if (!player || !actor) throw Rune.invalidAction()
 
         player.score += 1
         player.latestTurnScore += 1
@@ -67,9 +74,9 @@ Rune.initLogic({
 
         game.currentTurn.animal = getRandomItem(animals)
         game.currentTurn.emotion = getRandomItem(emotions)
-        if (game.currentTurn.timerStartedAt) {
-          game.currentTurn.timerStartedAt += displayCorrectGuessFor
-        }
+
+        if (!game.currentTurn.timerStartedAt) throw Rune.invalidAction()
+        game.currentTurn.timerStartedAt += displayCorrectGuessFor
       }
     },
     nextRound: (_, { game }) => {
@@ -86,11 +93,10 @@ Rune.initLogic({
     },
   },
   events: {
-    playerLeft: (playerId, { game }) => {
-      game.players = game.players.filter((p) => p.id !== playerId)
+    playerLeft: (_, { game }) => {
       startGameCheck(game)
 
-      if (!game.currentTurn) return
+      if (!game.currentTurn) throw Rune.invalidAction()
 
       if (
         game.currentTurn.stage === "acting" ||
@@ -160,39 +166,3 @@ Rune.initLogic({
     }
   },
 })
-
-function getRandomItem<T>(items: readonly T[]) {
-  return items[Math.floor(Math.random() * items.length)]
-}
-
-function startGameCheck(game: GameState) {
-  if (game.gameStarted) return
-  if (game.players.some((player) => !player.readyToStart)) return
-
-  game.gameStarted = true
-
-  setActor(game, "first")
-  newTurn(game)
-}
-
-function isLastActor(game: GameState) {
-  const actorIndex = game.players.findIndex((player) => player.actor)
-  return actorIndex + 1 === game.players.length
-}
-
-function setActor(game: GameState, which: "first" | "next") {
-  const actorIndex = game.players.findIndex((player) => player.actor)
-  const nextActorIndex = which === "first" ? 0 : actorIndex + 1
-
-  if (~actorIndex) game.players[actorIndex].actor = false
-  game.players[nextActorIndex].actor = true
-}
-
-function newTurn(game: GameState) {
-  game.currentTurn = {
-    animal: getRandomItem(animals),
-    emotion: getRandomItem(emotions),
-    stage: "countdown",
-    timerStartedAt: Rune.gameTimeInSeconds(),
-  }
-}
